Find number of subarrays with max sum of elements
A Company went to a college campus for hiring purpose. Various candidates participated in the
recruitment exam conducted by the employer. There are 10 final candidates left who qualified for the
vacancy. Employer maintains the marks of the candidate in the first test. Given a number K (1 ≤ K ≤
10), employer want to know that in how many ways he can select different set of K candidates out of 10
candidates such that their sum of marks in first test is maximum. Two set of K candidates would called
different if at least one of the candidate is not common in both set.
Input/Output Specifications
Input Specification:
1) An array of size 10 corresponding to the marks of 10 candidates.
2) Value K (1 ≤ K ≤ 10)
Output Specifications:
Number of different sets of K candidates corresponding to maximum summation
Examples
Example:
Input:
1) Candidate Marks
Candidate Marks
C1 2
C2 5
C3 1
C4 2
C5 4
C6 1
C7 6
C8 5
C9 2
C10 2
This is the marks of 10 candidates in the first test and employer has to select the team of 6 candidates
such that sum of their marks in first test is maximum.
2) 6 candidates
Output: 6 possible sets